fix(core): correctly format `NO_RESOURCE_NAME` arns
Currently there's an erroneous null check on `components.resourceName`
passed to `formatArn`. If the value isn't null, we add a separator and
resourceName to the arn output. `resourceName` cannot be null
(in TypeScript it's `string | undefined`) and thus this check will
always succeed leading us to to add a `sep` even when
`NO_RESOURCE_NAME` is the format and `resourceName` is an empty string
or undefined.

This fix checks firstly `NO_RESOURCE_NAME` is NOT the format and
secondly a value exists in `resourceName` before adding the `sep` and
`resourceName` to the output.